backend/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart   # needed for file uploads
